remove unnecessary functions
